+++ src/bika/lims/content/attachment.py
@@ -22,8 +22,10 @@ from AccessControl import ClassSecurityInfo
 from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from bika.lims import logger
+from bika.lims.browser.fields import UIDReferenceField
 from bika.lims.browser.fields.uidreferencefield import get_backreferences
 from bika.lims.browser.widgets import DateTimeWidget
+from bika.lims.browser.widgets import ReferenceWidget
 from bika.lims.config import ATTACHMENT_REPORT_OPTIONS
 from bika.lims.config import PROJECTNAME
 from bika.lims.content.bikaschema import BikaSchema
@@ -34,8 +36,6 @@ from plone.app.blob.field import FileField
 from Products.Archetypes.atapi import BaseFolder
 from Products.Archetypes.atapi import DateTimeField
 from Products.Archetypes.atapi import FileWidget
-from Products.Archetypes.atapi import ReferenceField
-from Products.Archetypes.atapi import ReferenceWidget
 from Products.Archetypes.atapi import Schema
 from Products.Archetypes.atapi import SelectionWidget
 from Products.Archetypes.atapi import StringField
@@ -52,11 +52,10 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         "AttachmentType",
         required=0,
         allowed_types=("AttachmentType",),
-        relationship="AttachmentAttachmentType",
         widget=ReferenceWidget(
             label=_("Attachment Type"),
         ),