+++ src/bika/lims/content/instrumentcertification.py
@@ -23,6 +23,7 @@ import math
 from AccessControl import ClassSecurityInfo
 from bika.lims import bikaMessageFactory as _
 from bika.lims import logger
+from bika.lims.browser.fields import UIDReferenceField
 from bika.lims.browser.widgets import ComboBoxWidget
 from bika.lims.browser.widgets import ReferenceWidget
 from bika.lims.config import PROJECTNAME
@@ -38,7 +39,6 @@ from Products.Archetypes.atapi import ComputedWidget
 from Products.Archetypes.atapi import DateTimeField
 from Products.Archetypes.atapi import DisplayList
 from Products.Archetypes.atapi import FileWidget
-from Products.Archetypes.atapi import ReferenceField
 from Products.Archetypes.atapi import Schema
 from Products.Archetypes.atapi import StringField
 from Products.Archetypes.atapi import StringWidget
@@ -59,10 +59,9 @@ schema = BikaSchema.copy() + Schema((
         )
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'Instrument',
         allowed_types=('Instrument',),
-        relationship='InstrumentCertificationInstrument',
         widget=StringWidget(
             visible=False,
         )
@@ -140,13 +139,11 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'Preparator',
         vocabulary='getLabContacts',
         allowed_types=('LabContact',),
-        relationship='LabContactInstrumentCertificatePreparator',
         widget=ReferenceWidget(
-            checkbox_bound=0,
             label=_("Prepared by"),
             description=_("The person at the supplier who prepared the certificate"),
             size=30,
@@ -160,13 +157,11 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         'Validator',
         vocabulary='getLabContacts',
         allowed_types=('LabContact',),
-        relationship='LabContactInstrumentCertificateValidator',
         widget=ReferenceWidget(
-            checkbox_bound=0,
             label=_("Approved by"),
             description=_("The person at the supplier who approved the certificate"),
             size=30,