+++ src/bika/lims/content/worksheettemplate.py
@@ -18,10 +18,11 @@
 # Copyright 2018-2021 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
-import sys
 from AccessControl import ClassSecurityInfo
 from bika.lims import api
 from bika.lims import bikaMessageFactory as _
+from bika.lims.browser.fields import UIDReferenceField
+from bika.lims.browser.widgets import ReferenceWidget
 from bika.lims.browser.widgets import ServicesWidget
 from bika.lims.browser.widgets import WorksheetTemplateLayoutWidget
 from bika.lims.config import ANALYSIS_TYPES
@@ -34,13 +35,10 @@ from Products.Archetypes.public import BaseContent
 from Products.Archetypes.public import BooleanField
 from Products.Archetypes.public import BooleanWidget
 from Products.Archetypes.public import DisplayList
-from Products.Archetypes.public import ReferenceField
-from Products.Archetypes.public import ReferenceWidget
 from Products.Archetypes.public import registerType
 from Products.Archetypes.public import Schema
 from Products.Archetypes.public import StringField
 from Products.Archetypes.public import StringWidget
-from Products.Archetypes.references import HoldingReference
 from senaite.core.browser.fields.records import RecordsField
 from senaite.core.catalog import SETUP_CATALOG
 from senaite.core.p3compat import cmp
@@ -74,14 +72,12 @@ schema = BikaSchema.copy() + Schema((
         )
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         "Service",
         schemata="Analyses",
         required=0,
         multiValued=1,
         allowed_types=("AnalysisService",),
-        relationship="WorksheetTemplateAnalysisService",
-        referenceClass=HoldingReference,
         widget=ServicesWidget(
             label=_("Analysis Service"),
             description=_(
@@ -90,17 +86,12 @@ schema = BikaSchema.copy() + Schema((
         )
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         "RestrictToMethod",
         schemata="Description",
-        required=0,
-        vocabulary_display_path_bound=sys.maxint,
         vocabulary="_getMethodsVoc",
         allowed_types=("Method",),
-        relationship="WorksheetTemplateMethod",
-        referenceClass=HoldingReference,
         widget=ReferenceWidget(
-            checkbox_bound=0,
             label=_("Method"),
             description=_(
                 "Restrict the available analysis services and instruments"
@@ -108,24 +99,21 @@ schema = BikaSchema.copy() + Schema((
                 " In order to apply this change to the services list, you "
                 "should save the change first."
             ),
+            showOn=True,
         ),
     ),
 
-    ReferenceField(
+    UIDReferenceField(
         "Instrument",
         schemata="Description",
-        required=0,
-        vocabulary_display_path_bound=sys.maxint,
         vocabulary="getInstrumentsDisplayList",
         allowed_types=("Instrument",),
-        relationship="WorksheetTemplateInstrument",
-        referenceClass=HoldingReference,
         widget=ReferenceWidget(
-            checkbox_bound=0,
             label=_("Instrument"),
             description=_(
                 "Select the preferred instrument"
             ),
+            showOn=True,
         ),
     ),